Fixed textarea issue
