+++ src/frontend/src/components/public/PIAFormTabs/review/MPOReviewSection.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { IPiaForm } from '../../../../types/interfaces/pia-form.interface';
-import { ReviewCard } from './ReviewCard';
+import { ReviewCard } from './ReviewCard/ReviewCard';
 import { getUserPrivileges } from '../../../../utils/statusList/common';
 import { getGUID } from '../../../../utils/user';
 import messages from './messages';
@@ -32,6 +32,7 @@ export const MPOReviewSection = (props: IMPOReviewProps) => {
     !printPreview && privilegedToEdit && !acknowledged,
   );
 
+  // Make read-only when canEditReview is false.
   const canEditReview =
     reviewGuid === userGUID && !printPreview && privilegedToEdit;
